+++ src/routers/round_robin_router.rs
@@ -115,7 +115,7 @@ where
 
 impl<A> RoundRobinRouter<A>
 where
-    A: Actor,
+    A: Actor + 'static,
 {
     pub fn new(stop_on_system_stop: bool, stop_on_empty_targets: bool) -> Self {
         Self {
@@ -130,7 +130,7 @@ where
 
 impl<A> Actor for RoundRobinRouter<A>
 where
-    A: Actor,
+    A: Actor + 'static,
 {
     fn on_system_stop(
         &mut self,
@@ -152,7 +152,7 @@ where
 
 impl<A> Handler<AddActorMessage<A>> for RoundRobinRouter<A>
 where
-    A: Actor,
+    A: Actor + 'static,
 {
     fn handle(
         &mut self,
@@ -167,7 +167,7 @@ where
 
 impl<A> Handler<RemoveActorMessage<A>> for RoundRobinRouter<A>
 where
-    A: Actor,
+    A: Actor + 'static,
 {
     fn handle(
         &mut self,